backend/requirements.txt
fastapi==0.111.0
torch==2.3.1
pillow==10.3.0
qdrant-client==1.9.2
ftfy==6.2.0
elasticsearch==8.14.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
requests==2.32.3
meilisearch==0.31.3